requirements.txt
Django==4.0.4
psycopg2-binary==2.9.3
djangorestframework==3.13.1
requests==2.27.1
